bank-account: move from settings to account
