+++ packages/pilot/src/containers/Settings/User/AccountInfoTab/CompanyGeneralForm.js
@@ -40,7 +40,7 @@ const CompanyGeneralForm = ({
           <Col palm={12} tablet={6} desk={6} tv={4}>
             <FormInput
               disabled
-              label={t('pages.settings.user.profile_info.general.form.name')}
+              label={t('pages.settings.user.account_info.general.form.name')}
               name="name"
               type="text"
             />
@@ -48,7 +48,7 @@ const CompanyGeneralForm = ({
           <Col palm={12} tablet={12} desk={6} tv={4}>
             <FormInput
               disabled
-              label={t('pages.settings.user.profile_info.general.form.full_name')}
+              label={t('pages.settings.user.account_info.general.form.full_name')}
               name="fullName"
               type="text"
             />
@@ -56,7 +56,7 @@ const CompanyGeneralForm = ({
           <Col palm={12} tablet={12} desk={3} tv={2}>
             <FormInput
               disabled
-              label={t('pages.settings.user.profile_info.general.form.cnpj')}
+              label={t('pages.settings.user.account_info.general.form.cnpj')}
               name="cnpj"
               type="text"
             />
@@ -64,7 +64,7 @@ const CompanyGeneralForm = ({
           <Col palm={12} tablet={12} desk={5} tv={5}>
             <FormInput
               disabled
-              label={t('pages.settings.user.profile_info.general.form.site_url')}
+              label={t('pages.settings.user.account_info.general.form.site_url')}
               name="siteUrl"
               type="text"
             />
